pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-eco-llm-inference-routing"
version = "0.1.0"
description = "Eco-LLM Inference Routing environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.8.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
package-dir = { "server" = "server" }
